FIX: now datastores and vnets are parsed correctly. Included more log lines. Included create_users function.

+++ ncm-opennebula/src/main/perl/opennebula.pm
@@ -53,10 +53,9 @@ sub process_template
     #my $tree = $config->getElement('/')->getTree();
     my $tpl = Template->new(INCLUDE_PATH => TEMPLATEPATH);
     #if (! $tpl->process($tt_rel, $tree, \$res)) {
-    if (! $tpl->process($tt_rel, $config, \$res)) {
+    if (! $tpl->process($tt_rel, { $tt_name => $config }, \$res)) {
             $self->error("TT processing of $tt_rel failed:", 
                                           $tpl->error());
-            $main::this_app->error("TT processing of $tt_rel failed:",$tpl->error());
             return;
     }
     return $res;
@@ -65,78 +64,109 @@ sub process_template
 # Create ONE resources
 sub create_something
 {
-    my ($self, $one, $data, $tt) = @_;
+    my ($self, $one, $tt, $data) = @_;
     
     my $template = $self->process_template($data, $tt);
     my $name;
     if (!$template) {
-        $main::this_app->error("No template data found for $tt!");
+        $self->error("No template data found for $tt.");
         return;
     }
     if ($template =~ m/^NAME\s+=\s+(?:"|')(.*?)(?:"|')\s*$/m) {
         $name = $1;
+        $self->verbose("Found template NAME:$name $tt resource.");
     } else {
-        $self->error("Template NAME not found!");
-        $main::this_app->error("Template NAME not found!");
+        $self->error("Template NAME not found.");
         return;
     }
     my $method = "create_$tt";
     my $new = $one->$method($template);
+    $self->info("It was created a new $name $tt resource.");
 }
 
 sub manage_something
 {
-    my ($self, $one, $resources, $type) = @_;
+    my ($self, $one, $type, $resources) = @_;
 
     if (!$resources) {
-        $main::this_app->error("No $type resources found!");
+        $self->error("No $type resources found.");
     } else {
-        $main::this_app->info("Found new $type resources...");
+        $self->info("Found new $type resources.");
     }
 
     if (($type eq "kvm") or ($type eq "xen")) {
-        $self->manage_hosts($one, $resources, $type);
+        $self->manage_hosts($one, $type, $resources);
         return;
     }
 
-    $main::this_app->info("Removing old ${type}/s...");
+    if ($type eq "user") {
+        $self->manage_users($one, $resources);
+        return;
+    }
+
+    $self->info("Removing old ${type}/s");
     my $method = "get_${type}s";
     my @existres = $one->$method(qr{^.*$});
     foreach my $oldresource (@existres) {
         $oldresource->delete();
     }
 
-    $main::this_app->info("Creating new ${type}/s...");
+    $self->info("Creating new ${type}/s");
     foreach my $newresource (@$resources) {
-        $self->create_something($one, $newresource, $type);
+        $self->create_something($one, $type, $newresource);
     }
 }
 
 
 sub manage_hosts
 {
-    my ($self, $one, @hosts, $type) = @_;
+    my ($self, $one, $type, $hosts) = @_;
 
-    $self->info("Removing old hosts...");
-    $main::this_app->info("Removing old hosts...");
-
+    $self->info("Removing old $type hosts.");
     my @existhost = $one->get_hosts(qr{^.*$});
     foreach my $t (@existhost) {
         $t->delete();
     }
 
-    $self->info("Creating new hosts...");
-    $main::this_app->info("Creating new hosts...");
-    foreach my $host (@hosts) {
+    foreach my $host (@$hosts) {
         my %host_options = (
             'name'    => $host, 
             'im_mad'  => $type, 
             'vmm_mad' => $type, 
             'vnm_mad' => "dummy"
         );
+        $self->verbose("Creating new $type host $host.");
         $one->create_host(%host_options);
     }
 }
+
+sub manage_users
+{
+    my ($self, $one, $users) = @_;
+
+    $self->info("Removing old regular users.");
+    my @exitsuser = $one->get_users(qr{^.*$});
+    foreach my $t (@exitsuser) {
+        $t->delete();
+    }
+
+    foreach my $user (@$users) {
+        if ($user->{user} && $user->{password}) {
+            $self->info("Creating new user: ", $user->{user});
+            $one->create_user($user->{user}, $user->{password}, "core");
+        }
+        else {
+            $self->error("No user or password info available.");
+        }
+    }
+
+}
+
+# DEBUG only (can't get the output in unittests otherwise)
+sub error { shift; $main::this_app->error(@_); };
+sub debug { shift; $main::this_app->debug(@_); };
+sub info { shift; $main::this_app->info(@_); };
+sub verbose { shift; $main::this_app->verbose(@_); };
 
 # Configure basic ONE resources
 sub Configure
@@ -147,27 +177,25 @@ sub Configure
     my $base = "/software/components/opennebula";
     my $tree = $config->getElement($base)->getTree();
 
-    $self->error("error");
     # Connect to ONE RPC
     my $one = $self->make_one($tree->{rpc});
     if (! $one ) {
         $self->error("No ONE instance created.");
-        $main::this_app->error("No ONE instance created.");
         return 0;
     };
 
     # Add/remove VNETs
-    $self->manage_something($one, $tree->{vnets}, "vnet");
+    $self->manage_something($one, "vnet", $tree->{vnets});
 
     # Add/remove datastores
-    $self->manage_something($one, $tree->{datastores}, "datastore");
+    $self->manage_something($one, "datastore", $tree->{datastores});
 
     # Add/remove KVM hosts
     my $hypervisor = "kvm";
-    $self->manage_something($one, $tree->{hosts}, $hypervisor);
+    $self->manage_something($one, $hypervisor, $tree->{hosts});
 
     # Add/remove regular users
-    #$self->manage_something($one, $tree->{users}, "user");
+    $self->manage_something($one, "user", $tree->{users});
 
     return 1;
 }
